update(refactor): use base classes for settings and main pages
