+++ cypress/e2e/02-friends.cy.ts
@@ -1,6 +1,6 @@
-import { authNewAccount } from "./PageObjects/AuthNewAccount";
-import { chatsMainPage } from "./PageObjects/ChatsMain";
-import { loginPinPage } from "./PageObjects/LoginPin";
+import authNewAccount from "./PageObjects/AuthNewAccount";
+import chatsMainPage from "./PageObjects/ChatsMain";
+import loginPinPage from "./PageObjects/LoginPin";
 
 describe("Friends Tests", () => {
   beforeEach(() => {
