+++ cypress/e2e/06-settings-inventory.cy.ts
@@ -1,8 +1,8 @@
-import { authNewAccount } from "./PageObjects/AuthNewAccount";
-import { chatsMainPage } from "./PageObjects/ChatsMain";
-import { loginPinPage } from "./PageObjects/LoginPin";
-import { settingsInventory } from "./PageObjects/Settings/SettingsInventory";
-import { settingsProfile } from "./PageObjects/Settings/SettingsProfile";
+import authNewAccount from "./PageObjects/AuthNewAccount";
+import chatsMainPage from "./PageObjects/ChatsMain";
+import loginPinPage from "./PageObjects/LoginPin";
+import settingsInventory from "./PageObjects/Settings/SettingsInventory";
+import settingsProfile from "./PageObjects/Settings/SettingsProfile";
 
 describe("Settings - Inventory", () => {
   beforeEach(() => {
@@ -30,23 +30,14 @@ describe("Settings - Inventory", () => {
     cy.url().should("include", "/settings/inventory");
 
     // Fire inventory frame should not be equipped
-    settingsInventory
-      .getFrameByName("Fire")
-      .find("[data-cy='inventory-item-button']")
-      .find("p")
-      .should("have.text", "Equip");
+    settingsInventory.getFrameButtonText("Fire").should("have.text", "Equip");
 
     // Equip Fire inventory frame
-    settingsInventory
-      .getFrameByName("Fire")
-      .find("[data-cy='inventory-item-button']")
-      .click();
+    settingsInventory.clickOnFrameButton("Fire");
 
     // Fire inventory frame should be equipped
-    cy.getByTestAttr("inventory-item-name")
-      .contains("Fire")
-      .last()
-      .parent()
+    settingsInventory
+      .getFrameContainer("Fire")
       .should("have.class", "equipped");
 
     settingsInventory.profilePictureFrameName.should("have.text", "Fire");
@@ -54,43 +45,43 @@ describe("Settings - Inventory", () => {
       "have.text",
       "Profile Picture Frame",
     );
+
+    settingsInventory.buttonProfile.click();
+    settingsProfile.profileImageFrame
+      .should("exist")
+      .and("have.attr", "src", "/assets/frames/fire.png");
+
+    settingsProfile.buttonInventory.click();
+    ``;
+
     settingsInventory.profilePictureFrameUnequipButton
       .should("contain", "Unequip")
       .click();
 
     // Fire inventory frame should be equipped
-    cy.getByTestAttr("inventory-item-name")
-      .contains("Fire")
-      .last()
-      .parent()
+    settingsInventory
+      .getFrameContainer("Fire")
       .should("not.have.class", "equipped");
+
+    settingsInventory.buttonProfile.click();
+    settingsProfile.profileImageFrame.should("not.exist");
   });
 
   it("J3, J4, J5, J6 - Equipping and unequipping inventory item", () => {
     cy.url().should("include", "/settings/inventory");
 
     // Fire inventory frame should not be equipped
-    settingsInventory
-      .getFrameByName("Fire")
-      .find("[data-cy='inventory-item-button']")
-      .find("p")
-      .should("have.text", "Equip");
+    settingsInventory.getFrameButtonText("Fire").should("have.text", "Equip");
 
     // Equip Fire inventory frame
-    settingsInventory
-      .getFrameByName("Fire")
-      .find("[data-cy='inventory-item-button']")
-      .click();
+    settingsInventory.clickOnFrameButton("Fire");
 
     // Fire inventory frame should be equipped
-    cy.getByTestAttr("inventory-item-name")
-      .contains("Fire")
-      .last()
-      .parent()
+    settingsInventory
+      .getFrameContainer("Fire")
       .should("have.class", "equipped");
 
-    cy.get('[data-cy="inventory-frame"].equipped')
-      .find("[data-cy='inventory-item-button']")
+    settingsInventory.inventoryFrameEquippedButton
       .should("have.css", "background-color", "rgb(77, 77, 255)")
       .should("contain", "Equipped");
 
@@ -104,10 +95,8 @@ describe("Settings - Inventory", () => {
       .click();
 
     // Fire inventory frame should be equipped
-    cy.getByTestAttr("inventory-item-name")
-      .contains("Fire")
-      .last()
-      .parent()
+    settingsInventory
+      .getFrameContainer("Fire")
       .should("not.have.class", "equipped");
   });
 });
