+++ cypress/e2e/PageObjects/Settings/SettingsProfile.ts
@@ -1,49 +1,9 @@
-class SettingsProfile {
-  get buttonAbout() {
-    return cy.getByTestAttr("button-About");
-  }
+import SettingsBase from "./SettingsBase";
 
-  get buttonAccessibility() {
-    return cy.getByTestAttr("button-Accessibility");
+class SettingsProfile extends SettingsBase {
+  constructor() {
+    super();
   }
-
-  get buttonAudioAndVideo() {
-    return cy.getByTestAttr("button-Audio & Video");
-  }
-
-  get buttonCustomization() {
-    return cy.getByTestAttr("button-Customization");
-  }
-
-  get buttonExtensions() {
-    return cy.getByTestAttr("button-Extensions");
-  }
-
-  get buttonInventory() {
-    return cy.getByTestAttr("button-Inventory");
-  }
-
-  get buttonKeybinds() {
-    return cy.getByTestAttr("button-Keybinds");
-  }
-
-  get buttonLicenses() {
-    return cy.getByTestAttr("button-Licenses");
-  }
-
-  get buttonMessages() {
-    return cy.getByTestAttr("button-Messages");
-  }
-
-  get buttonNotifications() {
-    return cy.getByTestAttr("button-Notifications");
-  }
-
-  get buttonProfile() {
-    return cy.getByTestAttr("button-Profile");
-  }
-
-  // Getters from Settings Profile
 
   get inputSettingsProfileShortID() {
     return cy.getByTestAttr("input-settings-profile-short-id");
@@ -93,6 +53,10 @@ class SettingsProfile {
     return this.profilePictureUploadButton
       .parents(".profile-picture-container")
       .siblings("input");
+  }
+
+  get profileImageFrame() {
+    return cy.getByTestAttr("profile-image-frame");
   }
 
   get profilePicture() {
@@ -288,4 +252,4 @@ class SettingsProfile {
   }
 }
 
-export const settingsProfile: SettingsProfile = new SettingsProfile();
+export default new SettingsProfile();
